+++ CardShop/CardShop/Models/Store.cs
@@ -19,11 +19,11 @@ namespace CardShop.Models
             this.StoreInventories = new HashSet<StoreInventory>();
             this.Transactions = new HashSet<Transaction>();
         }
-    
+
         public int StoreId { get; set; }
         public int UserId { get; set; }
         public string Name { get; set; }
-        public decimal DiscountRate { get; set; }
+        public Double DiscountRate { get; set; }
     
         public virtual ICollection<StoreInventory> StoreInventories { get; set; }
         public virtual ICollection<Transaction> Transactions { get; set; }
